pyproject.toml
[project]
name = "travel-planner-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "ddgs>=9.11.4",
    "fastapi[standard]>=0.135.1",
    "geopy>=2.4.1",
    "google-adk==2.0.0a1",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-openrouter>=0.1.0",
    "langgraph>=1.1.3",
    "logging>=0.4.9.6",
    "sqlmodel>=0.0.37",
]
